test: improve testing with puppeteer, so it does not fail randomly
Testing with puppeteer failed (started to fail quite often lately)
because of some timing issues between installing extension, checking
for updates and loading it on the tested page.
We now wait for extension to be requested from test server, then wait
for updates to be requested and then delay some more before reloading
test page and (finally!) checking if extension modified it as planned.