test: remove external CIs
CirrusCI and AppVeyor are great for automated tests and builds,
and they offer service for free for open source stuff which is
amazing and much appreciated.

Unfortunately it takes time to keep setup up-to-date when bumping
dependencies and/or nodejs version, so at least for now, only GitHub's
built-in actions will be used.