test: update puppeteer container file, allow disabling sandbox for test
Rootful docker is such a drag... a lot of time wasted on trying to run
test inside container with GitHub Actions.

Node.js uses default 1000:1000 for "node" user in their container images,
but GitHub runs containers as 1001:1001 as described at:
https://github.com/actions/toolkit/issues/1889

Matching UID:GID helps with file access/creation problems, but then...
Chromium cannot run anyway, because container is run as a "node"
user, but GitHub assumes containers are run as a "root" (AFAIK).
"Regular" user does not have permissions necessary for Chrome/Chromium
to setup sandbox.
When container is run as "root", there are no problems with filesystem,
but then Chrome/Chromium panics about running as "root"...

It's quite possible that there is a way to somehow modify user permissions,
but i don't have any more time to look for that.

That's why we now:
- build container with additional UID:GID that matches the one used
  by GitHub Action runners;
- run container as that additional user;
- launch Chromium "the old way", i.e., "full mode" with disabled sandbox
  and other stuff when running inside container from GitHub Action.

Since it takes a lot of time to test anything in such environment,
and i do not use locally Docker any more, i'm dropping example of testing
with Docker from the README.